backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
requests==2.28.2
python-dotenv==1.0.0
pydantic==1.10.7
scikit-learn==1.2.2
joblib==1.2.0
python-multipart==0.0.5